Converted a big part of hero updates to store
