+++ src/components/builder-mode/warrior/WarbandWarrior.tsx
@@ -3,7 +3,7 @@ import Badge from "react-bootstrap/Badge";
 import Card from "react-bootstrap/Card";
 import Stack from "react-bootstrap/Stack";
 import { Unit } from "../../../types/unit.ts";
-import { MwfBadge } from "../../common/MwfBadge.tsx";
+import { MwfBadge } from "../../common/might-will-fate/MwfBadge.tsx";
 import { UnitProfilePicture } from "../../images/UnitProfilePicture.tsx";
 import { WarriorActions } from "./WarriorActions.tsx";
 import { WarriorOptionList } from "./WarriorOptionList.tsx";
