+++ src/types/roster.ts
@@ -3,7 +3,7 @@ import { Warband } from "./warband.ts";
 export type Roster = {
   version: string;
   warbands: Warband[];
-  leader_warband_num: number | null;
+  leader_warband_id: string | null;
   num_units: number;
   bow_count: number;
   points: number;
